# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Content and Project Management API",
    "description": "API for managing content, projects, categories, subcategories, areas and statuses, with history and access control.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "creator"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "subcategories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Subcategory" }
          }
        }
      },
      "Subcategory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "categoryId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "category": { "$ref": "#/components/schemas/Category" }
        }
      },
      "Content": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string", "maxLength": 200 },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["draft", "published", "archived"], "default": "draft" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "authorId": { "type": "integer" },
          "lastUpdatedById": { "type": "integer", "nullable": true },
          "categoryId": { "type": "integer", "nullable": true },
          "subcategoryId": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "author": { "$ref": "#/components/schemas/User" },
          "lastUpdatedBy": { "$ref": "#/components/schemas/User" },
          "category": { "$ref": "#/components/schemas/Category" },
          "subcategory": { "$ref": "#/components/schemas/Subcategory" }
        }
      },
      "ContentHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "action": { "type": "string", "enum": ["created", "updated", "status_changed", "deleted"] },
          "changes": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "previousData": { "type": "string", "nullable": true },
          "newData": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" },
          "content": { "$ref": "#/components/schemas/Content" }
        }
      },
      "CreateContentDto": {
        "type": "object",
        "required": ["title", "description", "tags"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "minLength": 1 },
          "tags": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["draft", "published", "archived"] },
          "categoryId": { "type": "integer" },
          "subcategoryId": { "type": "integer" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateContentDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "minLength": 1 },
          "tags": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["draft", "published", "archived"] },
          "categoryId": { "type": "integer" },
          "subcategoryId": { "type": "integer" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
...
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 }
        }
      },
      "RegisterDto": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "role": { "type": "string", "enum": ["admin", "creator"], "default": "creator" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": { "type": "string" },
          "error": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar nuevo usuario (solo administradores)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "No autorizado - Se requiere autenticación de administrador",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Prohibido - Se requiere rol de administrador",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/content": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener lista de contenidos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "createdAt", "updatedAt", "title"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contenidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Content" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contenido"],
        "summary": "Crear nuevo contenido",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateContentDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contenido creado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Content" }
              }
            }
          }
        }
      }
    },
    "/content/{id}": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener contenido por ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Contenido encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Content" }
              }
            }
          },
          "404": {
            "description": "Contenido no encontrado"
          }
        }
      },
      "patch": {
        "tags": ["Contenido"],
        "summary": "Actualizar contenido",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateContentDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contenido actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Content" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contenido"],
        "summary": "Eliminar contenido",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Contenido eliminado exitosamente"
          }
        }
      }
    },
    "/content/{id}/history": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener historial de un contenido",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial del contenido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContentHistory" }
                }
              }
            }
          }
        }
      }
    },
    "/content/history/user/{userId}": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener historial de un usuario (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContentHistory" }
                }
              }
            }
          }
        }
      }
    },
    "/content/history/my": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener mi historial de contenido",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Mi historial de contenido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContentHistory" }
                }
              }
            }
          }
        }
      }
    },
    "/categories/public": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener categorías activas (público)",
        "parameters": [
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "name", "createdAt", "updatedAt"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías activas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener todas las categorías (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "description": "Filtrar por estado activo",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "name", "createdAt", "updatedAt"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categorías"],
        "summary": "Crear nueva categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener categoría por ID (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Categorías"],
        "summary": "Actualizar categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategoryDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Categorías"],
        "summary": "Eliminar categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría eliminada exitosamente"
          }
        }
      }
    },
    "/categories/{id}/toggle": {
      "patch": {
        "tags": ["Categorías"],
        "summary": "Cambiar estado de categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de categoría cambiado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
...
  }
} 